requirements.txt
anyio==4.11.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.1.1
python-telegram-bot==13.7
pytz==2025.2
requests==2.32.5
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tornado==6.5.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
